#50 refactor: hooks/use-event-listener・use-event-dispatcher を hooks/event/ へ再構成
同列ファイル間の相互 import・関心の未分離を解消するため、公開 hook
(useEventListener/useEventDispatcher/useEventPending) をそれぞれ
ディレクトリ化、共有 registry を _registries/index.ts へ集約
